Add Microsoft.Extensions.DependencyInjection.Abstractions package
Added Microsoft.Extensions.DependencyInjection.Abstractions package version 7.0.0 to the Replication.csproj file.

feat(Replication): Add ConfigureServices.cs

Added a new file ConfigureServices.cs to the Replication folder. The file contains a static class `ConfigureServices` with a method `AddReplicationServices` that adds replication services to the `IServiceCollection`.

Add ReplicationConnection.cs

Added a new file ReplicationConnection.cs to the Connections folder. The file contains a class `ReplicationConnection` that implements the `IReplicationConnection` interface. It has a constructor that takes a connection string and initializes a `LogicalReplicationConnection` object. The class also has methods `OpenAsync` and `StartAsync` for opening and starting replication with the connection.